refactor(wishlist): move admin route inline logic to controller (Fixes #651)
