+++ backend/controllers/wishlistController.js
@@ -839,4 +839,105 @@ const wishlistController = {
     }
 };
 
+// 1. Get any user's wishlist (Admin)
+exports.getAdminUserWishlist = async (req, res) => {
+    try {
+        const { safeInteger } = require("../utils/helpers"); // Require helper if not top-level
+        const db = require("../config/db");
+
+        const userId = safeInteger(req.params.userId);
+        if (!userId || userId < 1) {
+            return res.status(400).json({
+                success: false,
+                message: "Valid user ID is required",
+            });
+        }
+
+        const [rows] = await db.query(
+            `
+            SELECT 
+                p.id, 
+                p.name, 
+                p.price, 
+                p.image, 
+                p.brand, 
+                p.stock,
+                w.created_at as added_at,
+                u.name as user_name,
+                u.email as user_email
+            FROM wishlist_items w
+            JOIN products p ON w.product_id = p.id
+            JOIN users u ON w.user_id = u.id
+            WHERE w.user_id = ?
+            ORDER BY w.created_at DESC
+        `,
+            [userId],
+        );
+
+        return res.status(200).json({
+            success: true,
+            data: rows,
+        });
+    } catch (error) {
+        console.error("ADMIN GET WISHLIST ERROR:", error);
+        return res.status(500).json({
+            success: false,
+            message: "Failed to fetch user wishlist",
+        });
+    }
+};
+
+// 2. Get wishlist stats (Admin)
+exports.getWishlistStats = async (req, res) => {
+    try {
+        const db = require("../config/db");
+
+        // Get total wishlist items across all users
+        const [totalItems] = await db.query(
+            "SELECT COUNT(*) as total FROM wishlist_items",
+        );
+
+        // Get unique users with wishlist
+        const [uniqueUsers] = await db.query(
+            "SELECT COUNT(DISTINCT user_id) as users FROM wishlist_items",
+        );
+
+        // Get most wishlisted products
+        const [topProducts] = await db.query(`
+            SELECT p.id, p.name, COUNT(*) as wishlist_count
+            FROM wishlist_items w
+            JOIN products p ON w.product_id = p.id
+            GROUP BY p.id
+            ORDER BY wishlist_count DESC
+            LIMIT 10
+        `);
+
+        // Get recent activity
+        const [recentActivity] = await db.query(`
+            SELECT w.*, p.name as product_name, u.name as user_name
+            FROM wishlist_items w
+            JOIN products p ON w.product_id = p.id
+            JOIN users u ON w.user_id = u.id
+            ORDER BY w.created_at DESC
+            LIMIT 20
+        `);
+
+        return res.status(200).json({
+            success: true,
+            data: {
+                totalItems: totalItems[0]?.total || 0,
+                uniqueUsers: uniqueUsers[0]?.users || 0,
+                topProducts: topProducts,
+                recentActivity: recentActivity,
+            },
+        });
+    } catch (error) {
+        console.error("WISHLIST STATS ERROR:", error);
+        return res.status(500).json({
+            success: false,
+            message: "Failed to fetch wishlist stats",
+        });
+    }
+};
+
 module.exports = wishlistController;
